refactor: mvc 패턴의 패키지 구조로 패키지 구조 변경
Application 클래스의 역할을 CalculatorController에 위임하고
StringCalculator를 CalculatorService로 위임함
