+++ src/main/java/calculator/exception/CalculatorException.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.exception;
 
 public class CalculatorException extends IllegalArgumentException {
     public CalculatorException(String message) {
