+++ src/main/java/calculator/util/StringParser.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.util;
 
 import java.util.ArrayList;
 import java.util.Collections;
